# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto Nexus API
  description: |
    API REST do Projeto Nexus — Core System (Kernel) em Go sobre
    Arquitetura Microkernel: infraestrutura de autenticação, auditoria,
    outbox transacional e mensageria compartilhada por todo módulo de
    negócio (plug-in) registrado em `internal/modules/`. Todo endpoint
    autenticado espera um access token OIDC emitido pelo Keycloak (ou pelo
    login local, um caminho adicional — ver tag `auth`) como Bearer JWT; o
    backend o valida localmente contra um JWKS em cache (nunca com uma
    chamada por requisição ao Keycloak — ver §29 do documento de
    arquitetura).

    Toda resposta usa o mesmo envelope:
    `{"data": <payload|null>, "error": {"code","message"}|null, "meta": {...}|omitido}`.
    Erros nunca incluem stack trace ou outro detalhe interno.

    **RFC 7807 (Problem Details)**: um cliente que envia
    `Accept: application/problem+json` recebe as respostas de ERRO no
    formato RFC 7807 (`type` = `urn:nexus:error:<code>`, `title`,
    `status`, `detail`, `instance`, mais as extensões `code` e
    `request_id`). Qualquer outro `Accept` mantém o envelope acima.
    Respostas de sucesso não mudam. Ver ADR
    006-rfc7807-problem-details.md.

    **Idempotência**: qualquer endpoint de escrita (POST/PUT/PATCH) aceita
    opcionalmente o header `Idempotency-Key` (ver
    `#/components/parameters/IdempotencyKey`). Reenviar a mesma requisição
...

...
    de um provedor externo podem retornar 503
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/users/{id}:
    get:
      tags: [users]
      summary: Obtém um usuário pelo id
      operationId: getUser
      security:
        - bearerAuth: [users:read]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: O usuário.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Um access token OIDC emitido pelo Keycloak. Validado localmente
        (assinatura, emissor, audiência, expiração, algoritmo) contra um
        JWKS em cache.
...
  responses:
...
    Unauthorized:
      description: Access token ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          example:
            data: null
            error: { code: "UNAUTHORIZED", message: "invalid or expired access token" }
    Forbidden:
      description: Autenticado, mas sem o role/permissão necessário.
      content:
        application/json:
          example:
            data: null
            error: { code: "FORBIDDEN", message: "insufficient permission" }
    NotFound:
      description: O recurso não existe.
      content:
        application/json:
          example:
            data: null
            error: { code: "NOT_FOUND", message: "resource not found" }
...
  schemas:
    Envelope:
      type: object
      description: >
        O formato base de toda resposta. Operações bem-sucedidas restringem
        `data` a um tipo concreto via `allOf` no próprio schema da resposta;
        esta forma base documenta o contrato genérico {data, error, meta}.
      properties:
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ErrorBody"
        meta:
          type: object
          nullable: true
      required: [data, error]
    ErrorBody:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email: failed on \"required\""
      required: [code, message]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
...